Update account order, remove unused text modules
