--- app/graphql/customer-account/customer-order-query.ts
@@ -1,102 +1,0 @@
-// NOTE: https://shopify.dev/docs/api/customer/latest/queries/order
-export const CUSTOMER_ORDER_QUERY = `#graphql
-  fragment OrderMoney on MoneyV2 {
-    amount
-    currencyCode
-  }
-  fragment DiscountApplication on DiscountApplication {
-    ... on AutomaticDiscountApplication {
-      title
-    }
-    ... on DiscountCodeApplication {
-      code
-    }
-    value {
-      __typename
-      ... on MoneyV2 {
-        ...OrderMoney
-      }
-      ... on PricingPercentageValue {
-        percentage
-      }
-    }
-  }
-  fragment OrderLineItemFull on LineItem {
-    id
-    title
-    quantity
-    price {
-      ...OrderMoney
-    }
-    currentTotalPrice {
-      ...OrderMoney
-    }
-    totalPrice {
-      ...OrderMoney
-    }
-    discountAllocations {
-      allocatedAmount {
-        ...OrderMoney
-      }
-      discountApplication {
-        ...DiscountApplication
-      }
-    }
-    totalDiscount {
-      ...OrderMoney
-    }
-    image {
-      altText
-      height
-      url
-      id
-      width
-    }
-    variantTitle
-  }
-  fragment Order on Order {
-    id
-    name
-    statusPageUrl
-    processedAt
-    fulfillments(first: 1) {
-      nodes {
-        status
-      }
-    }
-    totalTax {
-      ...OrderMoney
-    }
-    totalPrice {
-      ...OrderMoney
-    }
-    subtotal {
-      ...OrderMoney
-    }
-    totalShipping {
-      ...OrderMoney
-    }
-    shippingAddress {
-      name
-      formatted(withName: true)
-      formattedArea
-    }
-    discountApplications(first: 100) {
-      nodes {
-        ...DiscountApplication
-      }
-    }
-    lineItems(first: 100) {
-      nodes {
-        ...OrderLineItemFull
-      }
-    }
-  }
-  query Order($orderId: ID!) {
-    order(id: $orderId) {
-      ... on Order {
-        ...Order
-      }
-    }
-  }
-` as const;
