--- app/modules/text.tsx
@@ -1,134 +1,0 @@
-import clsx from "clsx";
-import { formatText, missingClass } from "~/lib/utils";
-
-export function Text({
-  as: Component = "span",
-  className,
-  color = "default",
-  format,
-  size = "copy",
-  width = "default",
-  children,
-  ...props
-}: {
-  as?: React.ElementType;
-  className?: string;
-  color?: "default" | "primary" | "subtle" | "notice" | "contrast";
-  format?: boolean;
-  size?: "lead" | "copy" | "fine";
-  width?: "default" | "narrow" | "wide";
-  children: React.ReactNode;
-  [key: string]: any;
-}) {
-  const colors: Record<string, string> = {
-    default: "inherit",
-    primary: "text-body",
-    subtle: "text-body",
-    notice: "text-sale",
-    contrast: "text-body",
-  };
-
-  const sizes: Record<string, string> = {
-    lead: "text-lg font-medium",
-    copy: "",
-    fine: "text-xs subpixel-antialiased",
-  };
-
-  const widths: Record<string, string> = {
-    default: "max-w-prose",
-    narrow: "max-w-prose-narrow",
-    wide: "max-w-prose-wide",
-  };
-
-  const styles = clsx(
-    missingClass(className, "max-w-") && widths[width],
-    missingClass(className, "whitespace-") && "whitespace-pre-wrap",
-    missingClass(className, "text-") && colors[color],
-    sizes[size],
-    className,
-  );
-
-  return (
-    <Component {...props} className={styles}>
-      {format ? formatText(children) : children}
-    </Component>
-  );
-}
-
-export function Heading({
-  as: Component = "h2",
-  children,
-  className = "",
-  format,
-  size = "heading",
-  width = "default",
-  ...props
-}: {
-  as?: React.ElementType;
-  children: React.ReactNode;
-  format?: boolean;
-  size?: "display" | "heading" | "lead" | "copy";
-  width?: "default" | "narrow" | "wide";
-} & React.HTMLAttributes<HTMLHeadingElement>) {
-  const sizes = {
-    display: "font-bold text-4xl",
-    heading: "font-bold text-2xl leading-tight",
-    lead: "font-bold text-lg leading-snug",
-    copy: "font-medium text-base leading-normal",
-  };
-
-  const widths = {
-    default: "max-w-prose",
-    narrow: "max-w-prose-narrow",
-    wide: "max-w-prose-wide",
-  };
-
-  const styles = clsx(
-    missingClass(className, "whitespace-") && "whitespace-pre-wrap",
-    missingClass(className, "max-w-") && widths[width],
-    missingClass(className, "font-") && sizes[size],
-    className,
-  );
-
-  return (
-    <Component {...props} className={styles}>
-      {format ? formatText(children) : children}
-    </Component>
-  );
-}
-
-export function PageHeader({
-  children,
-  className,
-  heading,
-  variant = "default",
-  as: Component = "div",
-  ...props
-}: {
-  children?: React.ReactNode;
-  className?: string;
-  heading?: string;
-  variant?: "default" | "blogPost" | "allCollections";
-  [key: string]: any;
-}) {
-  const variants: Record<string, string> = {
-    default: "grid w-full gap-8 p-6 py-8 md:p-8 lg:p-12 justify-items-start",
-    blogPost:
-      "grid md:text-center w-full gap-4 p-6 py-8 md:p-8 lg:p-12 md:justify-items-center",
-    allCollections:
-      "flex justify-between items-baseline gap-8 p-6 md:p-8 lg:p-12",
-  };
-
-  const styles = clsx(variants[variant], className);
-
-  return (
-    <Component {...props} className={styles}>
-      {heading && (
-        <Heading as="h6" width="narrow" size="heading" className="inline-block">
-          {heading}
-        </Heading>
-      )}
-      {children}
-    </Component>
-  );
-}
