+++ app/routes/($locale).account.orders.$id.tsx
@@ -1,16 +1,11 @@
-import { Tag } from "@phosphor-icons/react";
-import { type MetaFunction, useLoaderData } from "@remix-run/react";
-import { Image, Money, flattenConnection } from "@shopify/hydrogen";
+import type { MetaFunction } from "@remix-run/react";
+import { flattenConnection } from "@shopify/hydrogen";
 import type { FulfillmentStatus } from "@shopify/hydrogen/customer-account-api-types";
 import { type LoaderFunctionArgs, json, redirect } from "@shopify/remix-oxygen";
-import clsx from "clsx";
-import type { OrderFragment } from "customer-accountapi.generated";
-import { Link } from "~/components/link";
-import { CUSTOMER_ORDER_QUERY } from "~/graphql/customer-account/customer-order-query";
-import { statusMessage } from "~/lib/utils";
-import { Text } from "~/modules/text";
+import type { OrderFragment, OrderQuery } from "customer-accountapi.generated";
+import { OrderDetails } from "~/components/account/order-details";
 
-export const meta: MetaFunction<typeof loader> = ({ data }) => {
+export let meta: MetaFunction<typeof loader> = ({ data }) => {
   return [{ title: `Order ${data?.order?.name}` }];
 };
 
@@ -19,15 +14,15 @@ export async function loader({ request, context, params }: LoaderFunctionArgs) {
     return redirect(params?.locale ? `${params.locale}/account` : "/account");
   }
 
-  const queryParams = new URL(request.url).searchParams;
-  const orderToken = queryParams.get("key");
+  let queryParams = new URL(request.url).searchParams;
+  let orderToken = queryParams.get("key");
 
   try {
-    const orderId = orderToken
+    let orderId = orderToken
       ? `gid://shopify/Order/${params.id}?key=${orderToken}`
       : `gid://shopify/Order/${params.id}`;
 
-    const { data, errors } = await context.customerAccount.query(
+    let { data, errors } = await context.customerAccount.query<OrderQuery>(
       CUSTOMER_ORDER_QUERY,
       { variables: { orderId } },
     );
@@ -36,22 +31,17 @@ export async function loader({ request, context, params }: LoaderFunctionArgs) {
       throw new Error("Order not found");
     }
 
-    const order: OrderFragment = data.order;
-
-    const lineItems = flattenConnection(order.lineItems);
-
-    const discountApplications = flattenConnection(order.discountApplications);
-
-    const firstDiscount = discountApplications[0]?.value;
-
-    const discountValue =
+    let order: OrderFragment = data.order;
+    let lineItems = flattenConnection(order.lineItems);
+    let discountApplications = flattenConnection(order.discountApplications);
+    let firstDiscount = discountApplications[0]?.value;
+    let discountValue =
       firstDiscount?.__typename === "MoneyV2" && firstDiscount;
-
-    const discountPercentage =
+    let discountPercentage =
       firstDiscount?.__typename === "PricingPercentageValue" &&
       firstDiscount?.percentage;
-    const fulfillments = flattenConnection(order.fulfillments);
-    const fulfillmentStatus =
+    let fulfillments = flattenConnection(order.fulfillments);
+    let fulfillmentStatus =
       fulfillments.length > 0
         ? fulfillments[0].status
         : ("OPEN" as FulfillmentStatus);
@@ -70,181 +60,107 @@ export async function loader({ request, context, params }: LoaderFunctionArgs) {
   }
 }
 
-export default function OrderRoute() {
-  const {
-    order,
-    lineItems,
-    discountValue,
-    discountPercentage,
-    fulfillmentStatus,
-  } = useLoaderData<typeof loader>();
-  let totalDiscount = 0;
-  lineItems.forEach((lineItem) => {
-    const itemDiscount = lineItem.discountAllocations.reduce(
-      (acc: number, curr) => {
-        return (acc += Number.parseFloat(curr.allocatedAmount.amount));
-      },
-      0,
-    );
-    totalDiscount += itemDiscount;
-  });
-  const totalDiscountMoney = {
-    amount: totalDiscount.toString(),
-    currencyCode: order.totalPrice?.currencyCode,
-  };
-  return (
-    <div className="max-w-screen-xl mx-auto px-4">
-      <div className="w-full p-4 sm:grid-cols-1 md:p-8 lg:p-12 lg:py-6">
-        <div className="flex flex-col gap-1 mb-8">
-          <div className="h4">Order Detail</div>
-          <Link to="/account">
-            <Text color="subtle">Return to My Account</Text>
-          </Link>
-        </div>
-        <div>
-          <p className="">Order No. {order.name}</p>
-          <p className="mt-2">
-            Placed on {new Date(order.processedAt).toDateString()}
-          </p>
-          <div className="grid grid-cols-1 md:grid-cols-3 mt-6">
-            <div className="space-y-6 col-span-2 md:pr-14">
-              {lineItems.map((lineItem) => {
-                let hasDiscount = false;
-                if (
-                  lineItem.currentTotalPrice?.amount !==
-                  lineItem.totalPrice?.amount
-                ) {
-                  hasDiscount = true;
-                }
-                return (
-                  <div className="flex gap-4" key={lineItem.id}>
-                    {lineItem?.image && (
-                      <div className="w-[120px] aspect-square shrink-0">
-                        <Image data={lineItem.image} width={120} height={120} />
-                      </div>
-                    )}
-                    <dl className="flex flex-col">
-                      <dt className="sr-only">Product</dt>
-                      <dd className="truncate">
-                        <div className="">{lineItem.title}</div>
-                        <div className="text-body-subtle text-sm">
-                          {lineItem.variantTitle}
-                        </div>
-                      </dd>
-                      <dt className="sr-only">Quantity</dt>
-                      <dd className="truncate mt-1 grow">
-                        x{lineItem.quantity}
-                      </dd>
-                      <dt className="sr-only">Discount</dt>
-                      <dd className="truncate flex gap-2 flex-wrap">
-                        {lineItem.discountAllocations.map((discount, index) => {
-                          const discountApp =
-                            discount.discountApplication as any;
-                          const discountTitle =
-                            discountApp?.title || discountApp?.code;
-                          return (
-                            <div
-                              className="text-body-subtle flex items-center gap-1 border border-[#B7B7B7] py-1 px-1.5 rounded-sm text-sm w-fit"
-                              key={index}
-                            >
-                              <Tag className="w-4 h-4" />
-                              <span>{discountTitle}</span>
-                              <div className="inline-flex">
-                                (<span>-</span>
-                                <Money data={discount.allocatedAmount} />)
-                              </div>
-                            </div>
-                          );
-                        })}
-                      </dd>
-                      <dt className="sr-only">Current Price</dt>
-                      <dd className="truncate flex gap-2 mt-2">
-                        {hasDiscount && (
-                          <span className="line-through text-body-subtle">
-                            <Money data={lineItem.totalPrice} />
-                          </span>
-                        )}
-                        <Money data={lineItem.currentTotalPrice} />
-                      </dd>
-                    </dl>
-                  </div>
-                );
-              })}
-              <hr className="border-t border-[#B7B7B7]" />
-              <div className="space-y-4 max-w-sm ml-auto">
-                <div className="flex justify-between gap-4 ">
-                  <span className="font-bold">Subtotal</span>
-                  <span>
-                    <Money data={order.subtotal} />
-                  </span>
-                </div>
-                <div className="flex justify-between gap-4">
-                  <span>Tax</span>
-                  <span>
-                    <Money data={order.totalTax} />
-                  </span>
-                </div>
-                <div className="flex justify-between gap-4">
-                  <span>Shipping</span>
-                  <span>
-                    <Money data={order.totalShipping} />
-                  </span>
-                </div>
-                <hr className="border-t border-[#B7B7B7] pb-2" />
-                <div className="flex justify-between gap-4">
-                  <span className="font-bold">Total</span>
-                  <span className="text-xl">
-                    <Money data={order.totalPrice} />
-                  </span>
-                </div>
-                <div className="flex justify-between gap-4">
-                  <div className="flex gap-1 items-center">
-                    <Tag className="w-4 h-4" />
-                    <span className="uppercase font-bold text-sm">
-                      Total savings
-                    </span>
-                  </div>
-                  <span>
-                    <Money data={totalDiscountMoney} />
-                  </span>
-                </div>
-              </div>
-            </div>
-            <div className="shrink-0 mt-4 pt-10 md:pt-0 md:m-0 md:border-none">
-              <div className="font-bold">Shipping Address</div>
-              {order?.shippingAddress ? (
-                <ul className="mt-3">
-                  <li>
-                    <Text>{order.shippingAddress.name}</Text>
-                  </li>
-                  {order?.shippingAddress?.formatted ? (
-                    order.shippingAddress.formatted.map((line: string) => (
-                      <li key={line}>
-                        <Text className="text-body-subtle">{line}</Text>
-                      </li>
-                    ))
-                  ) : (
-                    <></>
-                  )}
-                </ul>
-              ) : (
-                <p className="mt-3">No shipping address defined</p>
-              )}
-              <div className="font-bold mt-6">Status</div>
-              {fulfillmentStatus && (
-                <div
-                  className={clsx(
-                    "mt-3 px-2.5 py-1 text-sm inline-block w-auto",
-                    "border text-[#696662] border-[#696662]",
-                  )}
-                >
-                  <Text size="fine">{statusMessage(fulfillmentStatus!)}</Text>
-                </div>
-              )}
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
-  );
-}
+export default OrderDetails;
+
+// NOTE: https://shopify.dev/docs/api/customer/latest/queries/order
+const CUSTOMER_ORDER_QUERY = `#graphql
+  fragment OrderMoney on MoneyV2 {
+    amount
+    currencyCode
+  }
+  fragment DiscountApplication on DiscountApplication {
+    ... on AutomaticDiscountApplication {
+      title
+    }
+    ... on DiscountCodeApplication {
+      code
+    }
+    value {
+      __typename
+      ... on MoneyV2 {
+        ...OrderMoney
+      }
+      ... on PricingPercentageValue {
+        percentage
+      }
+    }
+  }
+  fragment OrderLineItemFull on LineItem {
+    id
+    title
+    quantity
+    price {
+      ...OrderMoney
+    }
+    currentTotalPrice {
+      ...OrderMoney
+    }
+    totalPrice {
+      ...OrderMoney
+    }
+    discountAllocations {
+      allocatedAmount {
+        ...OrderMoney
+      }
+      discountApplication {
+        ...DiscountApplication
+      }
+    }
+    totalDiscount {
+      ...OrderMoney
+    }
+    image {
+      altText
+      height
+      url
+      id
+      width
+    }
+    variantTitle
+  }
+  fragment Order on Order {
+    id
+    name
+    statusPageUrl
+    processedAt
+    fulfillments(first: 1) {
+      nodes {
+        status
+      }
+    }
+    totalTax {
+      ...OrderMoney
+    }
+    totalPrice {
+      ...OrderMoney
+    }
+    subtotal {
+      ...OrderMoney
+    }
+    totalShipping {
+      ...OrderMoney
+    }
+    shippingAddress {
+      name
+      formatted(withName: true)
+      formattedArea
+    }
+    discountApplications(first: 100) {
+      nodes {
+        ...DiscountApplication
+      }
+    }
+    lineItems(first: 100) {
+      nodes {
+        ...OrderLineItemFull
+      }
+    }
+  }
+  query Order($orderId: ID!) {
+    order(id: $orderId) {
+      ... on Order {
+        ...Order
+      }
+    }
+  }
+` as const;
